[System Tests] Add test cases for Kek rotation (#1391)
* Added rotation kek system test


* adding Experimental configuration


* adding experimental config fine


* added test case


* added new null checkings to avoid unexpected errors


* remove unused code


* fix format


* removed junit exception dependency


* added comment to skip AWS


* renamed Experimental to ExperimentalKmsConfig


* added comments and fix issues with kaf and kcat


* adding AssumptionViolationException class from Junit4


* replaced fluent logger in RecordEncryptionST


* fix pom


* added assertKekVersionWithinParcel


* using assumeTrue from junit instead of from assertj


* refactor assertKekRotationWithinParcel


* format


* remove commented code


* remove junit4 version from pom


* added suggestions


* moved random names to TestUtils and created list mutable


* supress thread sleep warnings


* added suggestions


---------

+++ kroxylicious-systemtests/src/test/java/io/kroxylicious/systemtests/RecordEncryptionST.java
@@ -7,10 +7,12 @@
 package io.kroxylicious.systemtests;
 
 import java.time.Duration;
+import java.util.ArrayList;
 import java.util.List;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.TestTemplate;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.slf4j.Logger;
@@ -26,6 +28,7 @@ import io.kroxylicious.systemtests.extensions.KroxyliciousExtension;
 import io.kroxylicious.systemtests.extensions.TestKubeKmsFacadeInvocationContextProvider;
 import io.kroxylicious.systemtests.installation.kroxylicious.Kroxylicious;
 import io.kroxylicious.systemtests.k8s.exception.KubeClusterException;
+import io.kroxylicious.systemtests.resources.kms.ExperimentalKmsConfig;
 import io.kroxylicious.systemtests.steps.KafkaSteps;
 import io.kroxylicious.systemtests.steps.KroxyliciousSteps;
 import io.kroxylicious.systemtests.templates.strimzi.KafkaNodePoolTemplates;
@@ -33,6 +36,7 @@ import io.kroxylicious.systemtests.templates.strimzi.KafkaTemplates;
 
 import static io.kroxylicious.systemtests.k8s.KubeClusterResource.kubeClient;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assumptions.assumeThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 @ExtendWith(KroxyliciousExtension.class)
@@ -61,18 +65,28 @@ class RecordEncryptionST extends AbstractST {
                 kafka);
     }
 
+    @BeforeEach
+    void beforeEach() {
+        bootstrap = null;
+        testKekManager = null;
+    }
+
     @AfterEach
     void afterEach(String namespace) {
         try {
-            LOGGER.atInfo().log("Deleting KEK...");
-            testKekManager.deleteKek("KEK_" + topicName);
+            if (testKekManager != null) {
+                LOGGER.atInfo().log("Deleting KEK...");
+                testKekManager.deleteKek("KEK_" + topicName);
+            }
         }
         catch (KubeClusterException e) {
             LOGGER.atError().setMessage("KEK deletion has not been successfully done: {}").addArgument(e).log();
             throw e;
         }
         finally {
-            KafkaSteps.deleteTopic(namespace, topicName, bootstrap);
+            if (bootstrap != null) {
+                KafkaSteps.deleteTopic(namespace, topicName, bootstrap);
+            }
         }
     }
 
@@ -83,21 +97,21 @@ class RecordEncryptionST extends AbstractST {
         int numberOfMessages = 1;
 
         // start Kroxylicious
-        LOGGER.atInfo().setMessage("Given Kroxylicious in {} namespace with {} replicas").addArgument(namespace).addArgument(1).log();
+        LOGGER.info("Given Kroxylicious in {} namespace with {} replicas", namespace, 1);
         Kroxylicious kroxylicious = new Kroxylicious(namespace);
         kroxylicious.deployPortPerBrokerPlainWithRecordEncryptionFilter(clusterName, 1, testKmsFacade);
         bootstrap = kroxylicious.getBootstrap();
 
-        LOGGER.atInfo().setMessage("And a kafka Topic named {}").addArgument(topicName).log();
+        LOGGER.info("And a kafka Topic named {}", topicName);
         KafkaSteps.createTopic(namespace, topicName, bootstrap, 1, 2);
 
-        LOGGER.atInfo().setMessage("When {} messages '{}' are sent to the topic '{}'").addArgument(numberOfMessages).addArgument(MESSAGE).addArgument(topicName).log();
+        LOGGER.info("When {} messages '{}' are sent to the topic '{}'", numberOfMessages, MESSAGE, topicName);
         KroxyliciousSteps.produceMessages(namespace, topicName, bootstrap, MESSAGE, numberOfMessages);
 
-        LOGGER.atInfo().setMessage("Then the messages are consumed").log();
+        LOGGER.info("Then the messages are consumed");
         List<ConsumerRecord> resultEncrypted = KroxyliciousSteps.consumeMessageFromKafkaCluster(namespace, topicName, clusterName,
                 Constants.KAFKA_DEFAULT_NAMESPACE, numberOfMessages, Duration.ofMinutes(2));
-        LOGGER.atInfo().setMessage("Received: {}").addArgument(resultEncrypted).log();
+        LOGGER.info("Received: {}", resultEncrypted);
 
         assertAll(
                 () -> assertThat(resultEncrypted.stream())
@@ -115,24 +129,144 @@ class RecordEncryptionST extends AbstractST {
         int numberOfMessages = 1;
 
         // start Kroxylicious
-        LOGGER.atInfo().setMessage("Given Kroxylicious in {} namespace with {} replicas").addArgument(namespace).addArgument(1).log();
+        LOGGER.info("Given Kroxylicious in {} namespace with {} replicas", namespace, 1);
         Kroxylicious kroxylicious = new Kroxylicious(namespace);
         kroxylicious.deployPortPerBrokerPlainWithRecordEncryptionFilter(clusterName, 1, testKmsFacade);
         bootstrap = kroxylicious.getBootstrap();
 
-        LOGGER.atInfo().setMessage("And a kafka Topic named {}").addArgument(topicName).log();
+        LOGGER.info("And a kafka Topic named {}", topicName);
         KafkaSteps.createTopic(namespace, topicName, bootstrap, 1, 2);
 
-        LOGGER.atInfo().setMessage("When {} messages '{}' are sent to the topic '{}'").addArgument(numberOfMessages).addArgument(MESSAGE).addArgument(topicName).log();
+        LOGGER.info("When {} messages '{}' are sent to the topic '{}'", numberOfMessages, MESSAGE, topicName);
         KroxyliciousSteps.produceMessages(namespace, topicName, bootstrap, MESSAGE, numberOfMessages);
 
-        LOGGER.atInfo().setMessage("Then the messages are consumed").log();
+        LOGGER.info("Then the messages are consumed");
         List<ConsumerRecord> result = KroxyliciousSteps.consumeMessages(namespace, topicName, bootstrap, numberOfMessages, Duration.ofMinutes(2));
-        LOGGER.atInfo().setMessage("Received: {}").addArgument(result).log();
+        LOGGER.info("Received: {}", result);
 
         assertThat(result).withFailMessage("expected messages have not been received!")
                 .extracting(ConsumerRecord::getValue)
                 .hasSize(numberOfMessages)
+                .allSatisfy(v -> assertThat(v).contains(MESSAGE));
+    }
+
+    @SuppressWarnings("java:S2925")
+    @TestTemplate
+    void ensureClusterHasEncryptedMessageWithRotatedKEK(String namespace, TestKmsFacade<?, ?, ?> testKmsFacade) {
+        // Skip AWS test execution because the ciphertext blob metadata to read the version of the KEK is not available anywhere
+        assumeThat(testKmsFacade.getKmsServiceClass().getSimpleName().toLowerCase().contains("vault")).isTrue();
+        testKekManager = testKmsFacade.getTestKekManager();
+        testKekManager.generateKek("KEK_" + topicName);
+        int numberOfMessages = 1;
+        ExperimentalKmsConfig experimentalKmsConfig = new ExperimentalKmsConfig(null, null, null, 5L);
+
+        // start Kroxylicious
+        LOGGER.info("Given Kroxylicious in {} namespace with {} replicas", namespace, 1);
+        Kroxylicious kroxylicious = new Kroxylicious(namespace);
+        kroxylicious.deployPortPerBrokerPlainWithRecordEncryptionFilter(clusterName, 1, testKmsFacade, experimentalKmsConfig);
+        bootstrap = kroxylicious.getBootstrap();
+
+        LOGGER.info("And a kafka Topic named {}", topicName);
+        KafkaSteps.createTopic(namespace, topicName, bootstrap, 1, 2);
+
+        LOGGER.info("When {} messages '{}' are sent to the topic '{}'", numberOfMessages, MESSAGE, topicName);
+        KroxyliciousSteps.produceMessages(namespace, topicName, bootstrap, MESSAGE, numberOfMessages);
+
+        LOGGER.info("Then the messages are consumed");
+        List<ConsumerRecord> resultEncrypted = KroxyliciousSteps.consumeMessageFromKafkaCluster(namespace, topicName, clusterName,
+                Constants.KAFKA_DEFAULT_NAMESPACE, numberOfMessages, Duration.ofMinutes(2));
+        LOGGER.info("Received: {}", resultEncrypted);
+
+        assertKekVersionWithinParcel(resultEncrypted, ":v1:", testKekManager);
+
+        LOGGER.info("When KEK is rotated");
+        testKekManager.rotateKek("KEK_" + topicName);
+
+        try {
+            Thread.sleep(5000);
+        }
+        catch (InterruptedException e) {
+            Thread.currentThread().interrupt();
+        }
+
+        LOGGER.info("And {} messages '{}' are sent to the topic '{}'", numberOfMessages, MESSAGE, topicName);
+        KroxyliciousSteps.produceMessages(namespace, topicName, bootstrap, MESSAGE, numberOfMessages);
+
+        LOGGER.info("Then the messages are consumed");
+        List<ConsumerRecord> resultEncryptedRotatedKek = KroxyliciousSteps.consumeMessageFromKafkaCluster(namespace, topicName, clusterName,
+                Constants.KAFKA_DEFAULT_NAMESPACE, numberOfMessages, Duration.ofMinutes(2));
+        LOGGER.info("Received: {}", resultEncryptedRotatedKek);
+
+        List<ConsumerRecord> finalEncryptedResults = new ArrayList<>(resultEncryptedRotatedKek);
+        finalEncryptedResults.removeAll(resultEncrypted);
+        assertKekVersionWithinParcel(finalEncryptedResults, ":v2:", testKekManager);
+    }
+
+    private void assertKekVersionWithinParcel(List<ConsumerRecord> consumerRecords, String expectedValue, TestKekManager testKekManager) {
+        assertThat(consumerRecords)
+                .withFailMessage("expected messages not received! Consumer records is empty")
+                .isNotEmpty();
+
+        assertThat(testKekManager.getClass().getSimpleName().toLowerCase())
+                .withFailMessage("Another KMS different from Vault is not currently supported!")
+                .contains("vault");
+
+        assertThat(consumerRecords.stream())
+                .withFailMessage(expectedValue + " is not contained in the ciphertext blob!")
+                .allMatch(r -> r.getValue().contains(expectedValue));
+    }
+
+    @SuppressWarnings("java:S2925")
+    @TestTemplate
+    void produceAndConsumeMessageWithRotatedKEK(String namespace, TestKmsFacade<?, ?, ?> testKmsFacade) {
+        testKekManager = testKmsFacade.getTestKekManager();
+        testKekManager.generateKek("KEK_" + topicName);
+        int numberOfMessages = 1;
+        boolean isVaultKms = testKmsFacade.getKmsServiceClass().getSimpleName().toLowerCase().contains("vault");
+        Long resolvedAliasExpireAfterWriteSeconds = isVaultKms ? null : 5L;
+        Long resolvedDekExpireAfterWriteSeconds = isVaultKms ? 5L : null;
+        ExperimentalKmsConfig experimentalKmsConfig = new ExperimentalKmsConfig(resolvedAliasExpireAfterWriteSeconds, null, null, resolvedDekExpireAfterWriteSeconds);
+
+        // start Kroxylicious
+        LOGGER.info("Given Kroxylicious in {} namespace with {} replicas", namespace, 1);
+        Kroxylicious kroxylicious = new Kroxylicious(namespace);
+        kroxylicious.deployPortPerBrokerPlainWithRecordEncryptionFilter(clusterName, 1, testKmsFacade, experimentalKmsConfig);
+        bootstrap = kroxylicious.getBootstrap();
+
+        LOGGER.info("And a kafka Topic named {}", topicName);
+        KafkaSteps.createTopic(namespace, topicName, bootstrap, 1, 2);
+
+        LOGGER.info("When {} messages '{}' are sent to the topic '{}'", numberOfMessages, MESSAGE, topicName);
+        KroxyliciousSteps.produceMessages(namespace, topicName, bootstrap, MESSAGE, numberOfMessages);
+
+        LOGGER.info("Then the messages are consumed");
+        List<ConsumerRecord> result = KroxyliciousSteps.consumeMessages(namespace, topicName, bootstrap, numberOfMessages, Duration.ofMinutes(2));
+        LOGGER.info("Received: {}", result);
+
+        assertThat(result).withFailMessage("expected messages have not been received!")
+                .extracting(ConsumerRecord::getValue)
+                .hasSize(numberOfMessages)
+                .allSatisfy(v -> assertThat(v).contains(MESSAGE));
+
+        LOGGER.info("When KEK is rotated");
+        testKekManager.rotateKek("KEK_" + topicName);
+
+        try {
+            Thread.sleep(5000);
+        }
+        catch (InterruptedException e) {
+            Thread.currentThread().interrupt();
+        }
+
+        LOGGER.info("And {} messages '{}' are sent to the topic '{}'", numberOfMessages, MESSAGE, topicName);
+        KroxyliciousSteps.produceMessages(namespace, topicName, bootstrap, MESSAGE, numberOfMessages);
+
+        LOGGER.info("Then the messages are consumed");
+        List<ConsumerRecord> resultRotatedKek = KroxyliciousSteps.consumeMessages(namespace, topicName, bootstrap, numberOfMessages, Duration.ofMinutes(2));
+        LOGGER.info("Received: {}", resultRotatedKek);
+
+        assertThat(resultRotatedKek).withFailMessage("expected messages have not been received!")
+                .extracting(ConsumerRecord::getValue)
                 .allSatisfy(v -> assertThat(v).contains(MESSAGE));
     }
 }
